Renames.
Former-commit-id: 0d0605bc65f04008b91155f88d51decce954b6f7
